imported_pip_requirements.txt
dj-database-url==0.5.0
Django==2.1
django-crispy-forms==1.7.2
django-filebrowser-no-grappelli==3.7.8
django-tinymce4-lite==1.7.5
jsmin==2.2.2
Pillow==6.1.0
pytz==2019.1
